requirements.txt
natto-py
numpy